cql_type_parser::raw_builder: Allow building types incrementally
Before this patch raw_builder would always start with an empty list of
user types. This means that every time a type is added to a keyspace,
every type in that keyspace needs to be recreated.

With this patch we pass a keyspace_metadata instead of just the
keyspace name and can construct new user types on top of previous
ones.

This will be used in the followup patch, where only new types are
created.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>