tools: toolchain: fix dbuild in interactive mode regression
Before ede1d248af0d65, running "tools/toolchain/dbuild -it -- bash" was
a nice way to play in the toolchain environment, for example to start
a debugger. But that commit caused containers to run in detached mode,
which is incompatible with interactive mode.

To restore the old behavior, detect that the user wants interactive mode,
and run the container in non-detached mode instead. Add the --rm flag
so the container is removed after execution (as it was before ede1d248af0d65).
Message-Id: <20190506175942.27361-1-avi@scylladb.com>