DynamoDB limits the size of hash keys to 2048 bytes, sort keys to 1024 bytes, and the entire item to 400 KB which therefore also limits the size of one attribute. This test checks that we can reach up to these limits, with binary keys and attributes.
The test does *not* check what happens once we exceed these
limits. In such a case, DynamoDB throws an error (I checked that
manually) but Alternator currently simply succeeds. If in the
future we decide to add artificial limits to Alternator as well,
we should add such tests as well.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>